refactor: Move ContactPhotos and GetContactPhotosResponse to separate file
